Organizes and renames files Does as title. Trying to make it the code base a little easier on the eyes
