+++ src/assets/styles/AuthPageStyle.css
