+++ src/services/routes.js
@@ -1,14 +1,14 @@
 import { createBrowserRouter } from "react-router-dom";
 
-import Index from "./pages";
-import SignUpPage from "./pages/SignUpPage";
-import LogInPage from "./pages/LogInPage";
-import CreateEventPage from "./pages/createEvent";
+import HomePage from "./pages/HomePage";
+import SignUpPage from "../pages/SignUpPage";
+import LogInPage from "../pages/LogInPage";
+import CreateEventPage from "../pages/CreateEventPage";
 
 const routes = createBrowserRouter([
     {
       path: "/",
-      element: <Index />,
+      element: <HomePage />,
     },
     {
       path: "/signup",
